refactor: Add fix to search functioanlity
