+++ src/Components/ArticleDetail/ArticleDetail.tsx
@@ -20,7 +20,7 @@ const ArticleDetail: React.FC<ArticleDetailProps> = ({ articles }) => {
 
   if (!article) return <div>Article not found. To go back to Main Page click on the Logo</div>;
 
-  console.log("Articles:", articles);
+  console.log("Articles:", articles); //dont forgert to delete
   console.log("Route ID:", id);
   console.log("Matched Article:", article);
 
@@ -29,18 +29,19 @@ const ArticleDetail: React.FC<ArticleDetailProps> = ({ articles }) => {
       <Link to="/" className="back-button">Back to Main Page</Link>
       <h1>{article.title}</h1>
       <img 
-    src={article.urlToImage || "path/to/default/image.jpg"} 
-    alt={article.title || "No Title"} 
-    className="detail-image" 
-    onError={(e) => { 
-        const imgElement = e.target as HTMLImageElement;
-        imgElement.onerror = null; 
-        imgElement.src = missingImg;
-    }}
-/>
+        src={article.urlToImage || "path/to/default/image.jpg"} 
+        alt={article.title || "No Title"} 
+        className="detail-image" 
+        onError={(e) => { 
+            const imgElement = e.target as HTMLImageElement;
+            imgElement.onerror = null; 
+            imgElement.src = missingImg;
+        }}
+      />
       <p><strong>Source:</strong> {article.source.name}</p>
       <p><strong>Date:</strong> {new Date(article.publishedAt).toLocaleDateString()}</p>
       <p>{article.content}</p>
+      <a href={article.url} target="_blank" rel="noopener noreferrer" className="article-original-link">Read full article on original source</a>
     </div>
   );
 }
