+++ src/Components/ArticleList/ArticleList.css
@@ -80,22 +80,59 @@ body {
     overflow: hidden;
 }
 
-.article-text-content p,
-.article-text-content h2 {
+.article-text-content p, .article-text-content h2 {
     word-wrap: break-word;
     margin: 5px 0;
     padding: 0 5px;
 }
 
 .search-container {
-    margin: 20px 0;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 5px;
+    background-color: #444;
+    margin-bottom: 20px;
+    border-radius: 5px;
+    box-shadow: 0px 2px 10px rgba(0, 0, 0, 0.1);
+    margin-top: -50px;
 }
 
-.search-input {
-    width: 100%;
-    padding: 10px;
-    border-radius: 4px;
-    border: 1px solid #e9e9e9;
+.search-input, #categorySelect {
+    border: none;
+    border-radius: 20px;
+    background-color: rgba(255, 255, 255, 0.1);
+    color: #fff;
+    transition: background-color 0.3s;
+    outline: none;
+    font-weight: bold;
+    flex: 1;
+    padding: 5px 10px;
+    font-size: 14px;
+    margin: 0 3px;
+}
+
+.search-button {
+    background-color: #e50914;
+    color: #fff;
+    border: none;
+    border-radius: 0 20px 20px 0;
+    padding: 5px 10px;
+    cursor: pointer;
+    transition: background-color 0.3s;
+    font-weight: bold;
+    outline: none;
+    display: flex;
+    align-items: center;
+}
+
+.search-button:hover {
+    background-color: #f40612;
+}
+
+.search-button::before {
+    content: "\1F50D";  
+    margin-right: 5px;
 }
 
 .article-detail-container {
@@ -141,7 +178,6 @@ body {
     margin: 0 auto;
 }
 
-/* Responsive Styles */
 @media (max-width: 1000px) {
     .article-link {
         flex: 1 1 calc(50% - 20px);
@@ -154,4 +190,70 @@ body {
         flex: 1 1 100%;
         max-width: 100%;
     }
+
+    .article-list-container {
+        margin-left: 0;
+    }
 }
+
+/* 
+.navbar-search {
+    display: flex;
+    align-items: center;
+}
+
+.navbar-search input {
+    padding: 5px 10px;
+    border: 1px solid #ddd;
+    outline: none;
+    margin-right: 10px;
+    transition: border 0.3s;
+    border-radius: 0 4px 4px 0;
+}
+
+.navbar-search input:focus {
+    border-color: #e50914;
+}
+
+.navbar-search button {
+    padding: 5px 15px;
+    background-color: #e50914;
+    color: #fff;
+    border: none;
+    border-radius: 4px;
+    cursor: pointer;
+    transition: background-color 0.2s;
+}
+
+.navbar-search button:hover {
+    background-color: #f40612;
+}
+
+.navbar-search select {
+    padding: 5px 10px;
+    border: 1px solid #ddd;
+    border-radius: 4px 0 0 0;
+    margin-right: 10px;
+    outline: none;
+    transition: border 0.3s;
+    cursor: pointer;
+}
+
+.navbar-search select:focus {
+    border-color: #e50914;
+}
+
+.navbar-search button[type="button"] {
+    background-color: transparent;
+    color: #e50914;
+    padding-left: 5px;
+    padding-right: 5px;
+    border: none;
+}
+
+@media (max-width: 768px) {
+    .navbar-search {
+        width: 100%;
+        justify-content: center;
+    }
+} */
